pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sat2graph-simplified"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "jvtubergen", email = "j.vantubergen@students.uu.nl" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
  "numpy",
  "opencv-python",
  "pillow",
  "rtree",
  "tensorflow",
  "requests", # gmaps-image dependency for API requests.
]

[project.urls]
Documentation = "https://github.com/jvtubergen/sat2graph-simplified#readme"
Issues = "https://github.com/jvtubergen/sat2graph-simplified/issues"
Source = "https://github.com/jvtubergen/sat2graph-simplified"

[tool.hatch.version]
path = "src/sat2graph-simplified/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/sat2graph-simplified tests}"

[tool.coverage.run]
source_pkgs = ["sat2graph-simplified", "tests"]
branch = true
parallel = true
omit = [
  "src/sat2graph-simplified/__about__.py",
]

[tool.coverage.paths]
sat2graph-simplified = ["src/sat2graph-simplified"]
tests = ["tests", "*/sat2graph-simplified/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
